requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
email_validator==2.1.1